+++ kernel/string.h
@@ -7,8 +7,8 @@
 //The destination buffer should contain space for at least len + 1 characters.
 void string_copy(char* dest, char* src, size_t len);
 
-char digit_to_char(uint digit);
+char digit_to_char(size_t digit);
 
-void uint_to_str(uint value, uint base, char* buf);
+void uint_to_str(size_t value, size_t base, char* buf);
 
 #endif
